Put properties type in same file as component, remove presets, and clean-up properties
