--- app/components/button/button.presets.ts
@@ -1,54 +1,0 @@
-import { ViewStyle, TextStyle } from "react-native"
-import { color, spacing } from "../../theme"
-
-/**
- * All text will start off looking like this.
- */
-const BASE_VIEW: ViewStyle = {
-  paddingVertical: spacing.small,
-  paddingHorizontal: spacing.small,
-  borderRadius: 4,
-  justifyContent: "center",
-  alignItems: "center",
-}
-
-const BASE_TEXT: TextStyle = {
-  paddingHorizontal: spacing.small,
-}
-
-/**
- * All the variations of text styling within the app.
- *
- * You want to customize these to whatever you need in your app.
- */
-export const viewPresets: Record<string, ViewStyle> = {
-  /**
-   * A smaller piece of secondard information.
-   */
-  primary: { ...BASE_VIEW, backgroundColor: color.primary } as ViewStyle,
-
-  /**
-   * A button without extras.
-   */
-  link: {
-    ...BASE_VIEW,
-    paddingHorizontal: 0,
-    paddingVertical: 0,
-    alignItems: "flex-start",
-  } as ViewStyle,
-}
-
-export const textPresets: Record<ButtonPresetNames, TextStyle> = {
-  primary: { ...BASE_TEXT, color: color.text } as TextStyle,
-  link: {
-    ...BASE_TEXT,
-    color: color.text,
-    paddingHorizontal: 0,
-    paddingVertical: 0,
-  } as TextStyle,
-}
-
-/**
- * A list of preset names.
- */
-export type ButtonPresetNames = keyof typeof viewPresets
