+++ app/components/button/button.tsx
@@ -1,30 +1,19 @@
 import * as React from "react"
-import { Text } from "../text/text"
-import { textPresets } from "./button.presets"
-import { ButtonProps } from "./button.props"
 import { Button as PaperButton } from "react-native-paper"
+
+// Having `extends PressableProps` does not work because `children` is not optional there.
+export interface ButtonProps {
+  title: string
+  onPress?: () => void
+  disabled?: boolean
+}
 
 /**
  * For your text displaying needs.
  *
  * This component is a HOC over the built-in React Native one.
  */
-export function Button(props: ButtonProps) {
-  // grab the props
-  const {
-    preset = "primary",
-    title: text,
-    textStyle: textStyleOverride,
-    children,
-    onPress,
-    disabled,
-  } = props
-
-  const textStyle = textPresets[preset] || textPresets.primary
-  const textStyles = [textStyle, textStyleOverride]
-
-  const content = children || <Text text={text} style={textStyles} />
-
+export function Button({ onPress, disabled, title }: ButtonProps) {
   // TODO Is setting `onMagicTap` necessary?
   return (
     <PaperButton
@@ -34,7 +23,7 @@ export function Button(props: ButtonProps) {
       onMagicTap={onPress}
       disabled={disabled}
     >
-      {content}
+      {title}
     </PaperButton>
   )
 }
