--- app/components/download-switch/download-switch.props.ts
@@ -1,5 +1,0 @@
-import { Track } from "../../models"
-
-export interface DownloadSwitchProps {
-  tracks: Track[]
-}
