--- app/components/icon/icon.props.ts
@@ -1,21 +1,0 @@
-import { ImageStyle, StyleProp, ViewStyle } from "react-native"
-import { IconTypes } from "./icons"
-
-export interface IconProps {
-  /**
-   * Style overrides for the icon image
-   */
-  style?: StyleProp<ImageStyle>
-
-  /**
-   * Style overrides for the icon container
-   */
-
-  containerStyle?: StyleProp<ViewStyle>
-
-  /**
-   * The name of the icon
-   */
-
-  icon: IconTypes
-}
