--- app/components/text/text.presets.ts
@@ -1,49 +1,0 @@
-import { TextStyle } from "react-native"
-import { color, typography } from "../../theme"
-import { scale } from "../../theme/scale"
-
-/**
- * All text will start off looking like this.
- */
-const BASE: TextStyle = {
-  fontFamily: typography.primary,
-  color: color.text,
-  fontSize: scale.medium,
-}
-
-/**
- * All the variations of text styling within the app.
- *
- * You want to customize these to whatever you need in your app.
- */
-export const presets = {
-  /**
-   * The default text styles.
-   */
-  default: BASE,
-
-  /**
-   * A bold version of the default text.
-   */
-  bold: { ...BASE, fontWeight: "bold" } as TextStyle,
-
-  /**
-   * Large headers.
-   */
-  header: { ...BASE, fontSize: scale.large, fontWeight: "bold" } as TextStyle,
-
-  /**
-   * Field labels that appear on forms above the inputs.
-   */
-  fieldLabel: { ...BASE, fontSize: scale.small, color: color.dim } as TextStyle,
-
-  /**
-   * A smaller piece of secondard information.
-   */
-  secondary: { ...BASE, fontSize: scale.tiny, color: color.dim } as TextStyle,
-}
-
-/**
- * A list of preset names.
- */
-export type TextPresets = keyof typeof presets
